requirements.txt
certifi==2016.2.28
future==0.15.2
python-telegram-bot==4.3
urllib3==1.16
youtube-dl==2016.6.27
